requirements.txt
python-dotenv
python-telegram-bot
requests
py-solc-x
web3
